renames ReviewTiles to RenderReviews, refacotred to tow components
